requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
numpy
python-multipart
gradio
transformers 
sentencepiece 
accelerate